Move core/TableClass.js to tables/TableClass.js
