+++ client/src/metadata/data_sources.js
@@ -8,9 +8,8 @@ import { glossaryEntryStore } from "src/models/glossary";
 import { lang, services_feature_flag } from "src/core/injected_build_constants";
 
 //circular dependency hack..
-import { Table } from "src/core/TableClass";
-
 import { rpb_link } from "src/rpb/rpb_link";
+import { Table } from "src/tables/TableClass";
 
 import data_source_text from "./data_sources.yaml";
 import freq_text from "./frequencies.yaml";
