+++ client/src/panels/panel_declarations/common_panel_utils.js
@@ -1,8 +1,7 @@
 import { PanelRegistry } from "src/panels/PanelRegistry";
 
-import { Table } from "src/core/TableClass";
-
 import { rpb_link, get_appropriate_rpb_subject } from "src/rpb/rpb_link";
+import { Table } from "src/tables/TableClass";
 
 const declare_panel = ({ panel_key, subject_types, panel_config_func }) => {
   if (!PanelRegistry.is_registered_panel_key(panel_key)) {
