+++ client/src/panels/panel_declarations/finances/vote_stat/in_year_estimates_split.js
@@ -10,8 +10,6 @@ import { businessConstants } from "src/models/businessConstants";
 import { formats } from "src/core/format";
 import { is_a11y_mode } from "src/core/injected_build_constants";
 
-import { filter_row_by_subj } from "src/core/TableClass";
-
 import { WrappedNivoBar } from "src/charts/wrapped_nivo/index";
 
 import {
@@ -19,6 +17,7 @@ import {
   secondaryColor,
   textColor,
 } from "src/style_constants/index";
+import { filter_row_by_subj } from "src/tables/TableClass";
 
 import { text_maker, TM } from "./vote_stat_text_provider";
